Information Value

This function calculates information value (IV) for multiple x variables.

iv(dt, y, x = NULL, positive = "bad|1", order = TRUE)

Arguments

dt
A data frame with both x (predictor/feature) and y (response/label) variables.

y
Name of y variable.

x
Name of x variables. Default is NULL. If x is NULL, then all variables except y are counted as x variables.

positive
Value of positive class, default is "bad|1".

order
Logical, default is TRUE. If it is TRUE, the output will descending order via iv.

Value

Information Value

Details

IV is a very useful concept for variable selection while developing credit scorecards. The formula for information value is shown below: $$IV = \sum(DistributionBad_{i} - DistributionGood_{i})*\ln(\frac{DistributionBad_{i}}{DistributionGood_{i}}).$$ The log component in information value is defined as weight of evidence (WOE), which is shown as $$WeightofEvidence = \ln(\frac{DistributionBad_{i}}{DistributionGood_{i}}).$$ The relationship between information value and predictive power is as follows: <0.02 (useless for prediction), 0.02 to 0.1 (Weak predictor), 0.1 to 0.3 (Medium predictor), 0.3 to 0.5 (Strong predictor) and >0.5 (Suspicious or too good to be true).

Examples

# Load German credit data
data(germancredit)

# information values
info_value = iv(germancredit, y = "creditability")

#> Warning: The positive value in "creditability" was replaced by 1 and negative value by 0.

info_value

#>                                                     variable   info_value
#>  1:                      status.of.existing.checking.account 6.660115e-01
#>  2:                                        duration.in.month 3.345035e-01
#>  3:                                           credit.history 2.932335e-01
#>  4:                                             age.in.years 2.596514e-01
#>  5:                                savings.account.and.bonds 1.960096e-01
#>  6:                                                  purpose 1.691951e-01
#>  7:                                                 property 1.126383e-01
#>  8:                                 present.employment.since 8.643363e-02
#>  9:                                                  housing 8.329343e-02
#> 10:                                  other.installment.plans 5.761454e-02
#> 11:                                           foreign.worker 4.387741e-02
#> 12:                                  personal.status.and.sex 4.268938e-02
#> 13:                                            credit.amount 3.895727e-02
#> 14:                              other.debtors.or.guarantors 3.201932e-02
#> 15:      installment.rate.in.percentage.of.disposable.income 2.632209e-02
#> 16:                  number.of.existing.credits.at.this.bank 1.326652e-02
#> 17:                                                      job 8.762766e-03
#> 18:                                                telephone 6.377605e-03
#> 19:                                  present.residence.since 3.588773e-03
#> 20: number.of.people.being.liable.to.provide.maintenance.for 4.339223e-05