One Hot Encoding

One-hot encoding on categorical variables and replace missing values. It is not needed when creating a standard scorecard model, but required in models that without doing woe transformation.

one_hot(dt, var_skip = NULL, var_encode = NULL, nacol_rm = FALSE,
  replace_na = NULL)

Arguments

dt
A data frame.

var_skip
Name of categorical variables that will skip for one-hot encoding. Defaults to NULL.

var_encode
Name of categorical variables to be one-hot encoded, Defaults to NULL. If it is NULL, then all categorical variables except in var_skip are counted.

nacol_rm
Logical. One-hot encoding on categorical variable contains missing values, whether to remove the column generated to indicate the presence of NAs. Defaults to FALSE.

replace_na
Replace missing values with a specified value such as -1, or the mean/median value for numeric variable and mode value for categorical variable. Defaults to NULL, which means no missing values will be replaced.

Value

A data frame

Examples

# load germancredit data
data(germancredit)

library(data.table)
dat = rbind(
  germancredit[, c(sample(20,3),21)],
  data.table(creditability=sample(c("good","bad"),10,replace=TRUE)),
  fill=TRUE)

# one hot encoding
## keep na columns from categorical variable
dat_onehot1 = one_hot(dat, var_skip = 'creditability', nacol_rm = FALSE) # default
str(dat_onehot1)

#> Classes ‘data.table’ and 'data.frame':	1010 obs. of  12 variables:
#>  $ creditability                                              : Factor w/ 2 levels "bad","good": 2 1 2 2 1 2 2 2 2 1 ...
#>  $ housing_NA                                                 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ housing_for free                                           : int  0 0 0 1 1 1 0 0 0 0 ...
#>  $ housing_own                                                : int  1 1 1 0 0 0 1 0 1 1 ...
#>  $ housing_rent                                               : int  0 0 0 0 0 0 0 1 0 0 ...
#>  $ personal.status.and.sex_NA                                 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ personal.status.and.sex_female : divorced/separated/married: int  0 1 0 0 0 0 0 0 1 0 ...
#>  $ personal.status.and.sex_male : married/widowed             : int  0 0 0 0 0 0 0 0 0 1 ...
#>  $ personal.status.and.sex_male : single                      : int  1 0 1 1 1 1 1 1 0 0 ...
#>  $ telephone_NA                                               : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ telephone_none                                             : int  0 1 1 1 1 0 1 0 1 1 ...
#>  $ telephone_yes, registered under the customers name         : int  1 0 0 0 0 1 0 1 0 0 ...
#>  - attr(*, ".internal.selfref")=<externalptr> 

## remove na columns from categorical variable
dat_onehot2 = one_hot(dat, var_skip = 'creditability', nacol_rm = TRUE)
str(dat_onehot2)

#> Classes ‘data.table’ and 'data.frame':	1010 obs. of  9 variables:
#>  $ creditability                                              : Factor w/ 2 levels "bad","good": 2 1 2 2 1 2 2 2 2 1 ...
#>  $ housing_for free                                           : int  0 0 0 1 1 1 0 0 0 0 ...
#>  $ housing_own                                                : int  1 1 1 0 0 0 1 0 1 1 ...
#>  $ housing_rent                                               : int  0 0 0 0 0 0 0 1 0 0 ...
#>  $ personal.status.and.sex_female : divorced/separated/married: int  0 1 0 0 0 0 0 0 1 0 ...
#>  $ personal.status.and.sex_male : married/widowed             : int  0 0 0 0 0 0 0 0 0 1 ...
#>  $ personal.status.and.sex_male : single                      : int  1 0 1 1 1 1 1 1 0 0 ...
#>  $ telephone_none                                             : int  0 1 1 1 1 0 1 0 1 1 ...
#>  $ telephone_yes, registered under the customers name         : int  1 0 0 0 0 1 0 1 0 0 ...
#>  - attr(*, ".internal.selfref")=<externalptr> 

## one hot and replace NAs
dat_onehot3 = one_hot(dat, var_skip = 'creditability', replace_na = -1)
str(dat_onehot3)

#> Classes ‘data.table’ and 'data.frame':	1010 obs. of  12 variables:
#>  $ creditability                                              : Factor w/ 2 levels "bad","good": 2 1 2 2 1 2 2 2 2 1 ...
#>  $ housing_NA                                                 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ housing_for free                                           : int  0 0 0 1 1 1 0 0 0 0 ...
#>  $ housing_own                                                : int  1 1 1 0 0 0 1 0 1 1 ...
#>  $ housing_rent                                               : int  0 0 0 0 0 0 0 1 0 0 ...
#>  $ personal.status.and.sex_NA                                 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ personal.status.and.sex_female : divorced/separated/married: int  0 1 0 0 0 0 0 0 1 0 ...
#>  $ personal.status.and.sex_male : married/widowed             : int  0 0 0 0 0 0 0 0 0 1 ...
#>  $ personal.status.and.sex_male : single                      : int  1 0 1 1 1 1 1 1 0 0 ...
#>  $ telephone_NA                                               : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ telephone_none                                             : int  0 1 1 1 1 0 1 0 1 1 ...
#>  $ telephone_yes, registered under the customers name         : int  1 0 0 0 0 1 0 1 0 0 ...
#>  - attr(*, ".internal.selfref")=<externalptr> 

# replace missing values only
## replace with -1
dat_repna1 = one_hot(dat, var_skip = names(dat), replace_na = -1)
## replace with median for numeric, and mode for categorical
dat_repna2 = one_hot(dat, var_skip = names(dat), replace_na = 'median')
## replace with to mean for numeric, and mode for categorical
dat_repna3 = one_hot(dat, var_skip = names(dat), replace_na = 'mean')