Creating a Scorecard

scorecard creates a scorecard based on the results from woebin and glm.

scorecard(bins, model, points0 = 600, odds0 = 1/19, pdo = 50,
  basepoints_eq0 = FALSE, digits = 0)

Arguments

bins

Binning information generated from woebin function.

model

A glm model object.

points0

Target points, default 600.

odds0

Target odds, default 1/19. Odds = p/(1-p).

pdo

Points to Double the Odds, default 50.

basepoints_eq0

Logical, Defaults to FALSE. If it is TRUE, the basepoints will equally distribute to each variable.

digits

The number of digits after the decimal point for points calculation. Default 0.

Value

A list of scorecard data frames

See also

Examples

# \donttest{
# load germancredit data
data("germancredit")

# filter variable via missing rate, iv, identical value rate
dt_sel = var_filter(germancredit, "creditability")
#> [INFO] filtering variables ... 

# woe binning ------
bins = woebin(dt_sel, "creditability")
#> [INFO] creating woe binning ... 
dt_woe = woebin_ply(dt_sel, bins)
#> [INFO] converting into woe values ... 

# glm ------
m = glm(creditability ~ ., family = binomial(), data = dt_woe)

# Select a formula-based model by AIC
m_step = step(m, direction="both", trace=FALSE)
m = eval(m_step$call)

# scorecard
# Example I # creat a scorecard
card = scorecard(bins, m)

xnames = sub('_woe', '', names(coef(m))[-1])
card2 = scorecard2(bins=bins, dt=germancredit, y='creditability', x=xnames)

# credit score
# Example I # only total score
score1 = scorecard_ply(germancredit, card)

# Example II # credit score for both total and each variable
score2 = scorecard_ply(germancredit, card, only_total_score = FALSE)
# }