Creating a Scorecard

scorecard2 creates a scorecard based on the results from woebin. It has the same function with scorecard, but without model object input.

scorecard2(bins, dt, y, x = NULL, points0 = 600, odds0 = 1/19,
  pdo = 50, basepoints_eq0 = FALSE, positive = "bad|1", ...)

Arguments

bins
Binning information generated from woebin function.

dt
A data frame with both x (predictor/feature) and y (response/label) variables.

y
Name of y variable.

x
Name of x variables. If it is NULL, then all variables in bins are used. Default is NULL.

points0
Target points, default 600.

odds0
Target odds, default 1/19. Odds = p/(1-p).

pdo
Points to Double the Odds, default 50.

basepoints_eq0
Logical, default is FALSE. If it is TRUE, the basepoints will equally distribute to each variable.

positive
Value of positive class, default "bad|1".

...
Additional parameters.

Value

A scorecard dataframe

See also

Examples

# NOT RUN {
# load germancredit data
data("germancredit")

# filter variable via missing rate, iv, identical value rate
dt_sel = var_filter(germancredit, "creditability")

# woe binning ------
bins = woebin(dt_sel, "creditability")
dt_woe = woebin_ply(dt_sel, bins)

# glm ------
m = glm(creditability ~ ., family = binomial(), data = dt_woe)
# summary(m)

# Select a formula-based model by AIC
m_step = step(m, direction="both", trace=FALSE)
m = eval(m_step$call)
# summary(m)

# predicted proability
# dt_pred = predict(m, type='response', dt_woe)

# performace
# ks & roc plot
# perf_eva(dt_woe$creditability, dt_pred)

# scorecard
# Example I # creat a scorecard
card = scorecard(bins, m)
card2 = scorecard2(bins=bins, dt=germancredit, y='creditability',
  x=c("status.of.existing.checking.account", "duration.in.month", "credit.history",
   "purpose", "credit.amount", "savings.account.and.bonds",
   "present.employment.since", "installment.rate.in.percentage.of.disposable.income",
   "personal.status.and.sex", "other.debtors.or.guarantors", "property",
   "age.in.years", "other.installment.plans", "housing"))

# credit score
# Example I # only total score
score1 = scorecard_ply(dt, card)

# Example II # credit score for both total and each variable
score2 = scorecard_ply(dt, card, only_total_score = F)
# }