WOE Binning

woebin generates optimal binning for numerical, factor and categorical variables using methods including tree-like segmentation or chi-square merge. woebin can also customizing breakpoints if the breaks_list was provided.

woebin(dt, y, x = NULL, breaks_list = NULL, special_values = NULL,
  min_perc_fine_bin = 0.02, min_perc_coarse_bin = 0.05, stop_limit = 0.1,
  max_num_bin = 8, positive = "bad|1", no_cores = NULL, print_step = 0L,
  method = "tree")

Arguments

dt
A data frame with both x (predictor/feature) and y (response/label) variables.

y
Name of y variable.

x
Name of x variables. Default is NULL. If x is NULL, then all variables except y are counted as x variables.

breaks_list
List of break points, default is NULL. If it is not NULL, variable binning will based on the provided breaks.

special_values
the values specified in special_values will be in separate bins. Default is NULL.

min_perc_fine_bin
The minimum percentage of initial binning class number over total. Accepted range: 0.01-0.2; default is 0.02, which means initial binning into 50 fine bins for continuous variables.

min_perc_coarse_bin
The minimum percentage of final binning class number over total. Accepted range: 0.01-0.2; default is 0.05.

stop_limit
Stop binning segmentation when information value gain ratio less than the stop_limit, or stop binning merge when the minimum of chi-square less than 'qchisq(1-stoplimit, 1)'. Accepted range: 0-0.5; default is 0.1.

max_num_bin
Integer. The maximum number of binning.

positive
Value of positive class, default "bad|1".

no_cores
Number of CPU cores for parallel computation. Defaults NULL. If no_cores is NULL, the no_cores will set as 1 if length of x variables less than 10, and will set as the number of all CPU cores if the length of x variables greater than or equal to 10.

print_step
A non-negative integer. Default is 1. If print_step>0, print variable names by each print_step-th iteration. If print_step=0 or no_cores>1, no message is print.

method
Optimal binning method, it should be "tree" or "chimerge". Default is "tree".

Value

Optimal or customized binning information.

See also

Examples

# load germancredit data
data(germancredit)

# Example I
# binning of two variables in germancredit dataset
# using tree method
bins2_tree = woebin(germancredit, y="creditability",
   x=c("credit.amount","housing"), method="tree")

#> Warning: The positive value in "creditability" was replaced by 1 and negative value by 0.

bins2_tree

#> $credit.amount
#>         variable         bin count count_distr good bad   badprob         woe
#> 1: credit.amount [-Inf,1400)   267       0.267  185  82 0.3071161  0.03366128
#> 2: credit.amount [1400,1800)   105       0.105   87  18 0.1714286 -0.72823850
#> 3: credit.amount [1800,4000)   382       0.382  287  95 0.2486911 -0.25830746
#> 4: credit.amount [4000,9200)   196       0.196  120  76 0.3877551  0.39053946
#> 5: credit.amount [9200, Inf)    50       0.050   21  29 0.5800000  1.17007125
#>          bin_iv  total_iv breaks is_special_values
#> 1: 0.0003045545 0.1812204   1400             FALSE
#> 2: 0.0468153322 0.1812204   1800             FALSE
#> 3: 0.0241086966 0.1812204   4000             FALSE
#> 4: 0.0319870413 0.1812204   9200             FALSE
#> 5: 0.0780047502 0.1812204    Inf             FALSE
#> 
#> $housing
#>    variable      bin count count_distr good bad   badprob        woe     bin_iv
#> 1:  housing     rent   179       0.179  109  70 0.3910615  0.4044452 0.03139265
#> 2:  housing      own   713       0.713  527 186 0.2608696 -0.1941560 0.02579501
#> 3:  housing for free   108       0.108   64  44 0.4074074  0.4726044 0.02610577
#>      total_iv   breaks is_special_values
#> 1: 0.08329343     rent             FALSE
#> 2: 0.08329343      own             FALSE
#> 3: 0.08329343 for free             FALSE
#> 

# NOT RUN {
# using chimerge method
bins2_chi = woebin(germancredit, y="creditability",
   x=c("credit.amount","housing"), method="chimerge")

# Example II
# binning of the germancredit dataset
bins_germ = woebin(germancredit, y = "creditability")
# converting bins_germ into a dataframe
# bins_germ_df = data.table::rbindlist(bins_germ)

# Example III
# customizing the breakpoints of binning
library(data.table)
dat = rbind(
  germancredit,
  data.table(creditability=sample(c("good","bad"),10,replace=TRUE)),
  fill=TRUE)

breaks_list = list(
  age.in.years = c(26, 35, 37, "Inf%,%missing"),
  housing = c("own", "for free%,%rent")
)

special_values = list(
  credit.amount = c(2600, 9960, "6850%,%missing"),
  purpose = c("education", "others%,%missing")
)

bins_cus_brk = woebin(dat, y="creditability",
  x=c("age.in.years","credit.amount","housing","purpose"),
  breaks_list=breaks_list, special_values=special_values)

# }