WOE Binning Adjustment

woebin_adj interactively adjust the binning breaks.

woebin_adj(dt, y, bins, breaks_list = NULL, save_breaks_list = NULL,
  adj_all_var = TRUE, special_values = NULL, method = "tree",
  count_distr_limit = 0.05, to = "breaks_list", ...)

Arguments

dt

A data frame.

y

Name of y variable.

bins

A list of data frames. Binning information generated from woebin.

breaks_list

List of break points, Defaults to NULL. If it is not NULL, variable binning will based on the provided breaks.

save_breaks_list

A string. The file name to save breaks_list. Defaults to None.

adj_all_var

Logical, whether to show variables have monotonic woe trends. Defaults to TRUE

special_values

The values specified in special_values will in separate bins. Defaults to NULL.

method

Optimal binning method, it should be "tree" or "chimerge". Defaults to "tree".

count_distr_limit

The minimum count distribution percentage. Accepted range: 0.01-0.2; Defaults to 0.05. This argument should be the same with woebin's.

to

Adjusting bins into breaks_list or bins_list. Defaults to breaks_list.

...

Additional parameters.

Value

A list of modified break points of each x variables.

See also

Examples

if (FALSE) {
# Load German credit data
data(germancredit)

# Example I
dt = germancredit[, c("creditability", "age.in.years", "credit.amount")]
bins = woebin(dt, y="creditability")
breaks_adj = woebin_adj(dt, y="creditability", bins)
bins_final = woebin(dt, y="creditability",
                    breaks_list=breaks_adj)

# Example II adjust two variables' breaks in brklst
binsII = woebin(germancredit, y="creditability", save_breaks_list = 'brklst')
brklst = source('brklst.R')$value
brklst_adj = woebin_adj(germancredit, "creditability", binsII[1:2], breaks_list = brklst)
}