German Credit Data

Credit data that classifies debtors described by a set of attributes as good or bad credit risks. See source link below for detailed information.

data(germancredit)

Format

A data frame with 21 variables (numeric and factors) and 1000 observations.

Source

https://archive.ics.uci.edu/ml/datasets/Statlog+(German+Credit+Data)

Examples

# load German credit data
data(germancredit)

# structure of germancredit
str(germancredit)

#> 'data.frame':	1000 obs. of  21 variables:
#>  $ status.of.existing.checking.account                     : Factor w/ 4 levels "... < 0 DM","0 <= ... < 200 DM",..: 1 2 4 1 1 4 4 2 4 2 ...
#>  $ duration.in.month                                       : num  6 48 12 42 24 36 24 36 12 30 ...
#>  $ credit.history                                          : Factor w/ 5 levels "no credits taken/ all credits paid back duly",..: 5 3 5 3 4 3 3 3 3 5 ...
#>  $ purpose                                                 : chr  "radio/television" "radio/television" "education" "furniture/equipment" ...
#>  $ credit.amount                                           : num  1169 5951 2096 7882 4870 ...
#>  $ savings.account.and.bonds                               : Factor w/ 5 levels "... < 100 DM",..: 5 1 1 1 1 5 3 1 4 1 ...
#>  $ present.employment.since                                : Factor w/ 5 levels "unemployed","... < 1 year",..: 5 3 4 4 3 3 5 3 4 1 ...
#>  $ installment.rate.in.percentage.of.disposable.income     : num  4 2 2 2 3 2 3 2 2 4 ...
#>  $ personal.status.and.sex                                 : Factor w/ 5 levels "male : divorced/separated",..: 3 2 3 3 3 3 3 3 2 4 ...
#>  $ other.debtors.or.guarantors                             : Factor w/ 3 levels "none","co-applicant",..: 1 1 1 3 1 1 1 1 1 1 ...
#>  $ present.residence.since                                 : num  4 2 3 4 4 4 4 2 4 2 ...
#>  $ property                                                : Factor w/ 4 levels "real estate",..: 1 1 1 2 4 4 2 3 1 3 ...
#>  $ age.in.years                                            : num  67 22 49 45 53 35 53 35 61 28 ...
#>  $ other.installment.plans                                 : Factor w/ 3 levels "bank","stores",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ housing                                                 : Factor w/ 3 levels "rent","own","for free": 2 2 2 3 3 3 2 1 2 2 ...
#>  $ number.of.existing.credits.at.this.bank                 : num  2 1 1 1 2 1 1 1 1 2 ...
#>  $ job                                                     : Factor w/ 4 levels "unemployed/ unskilled - non-resident",..: 3 3 2 3 3 2 3 4 2 4 ...
#>  $ number.of.people.being.liable.to.provide.maintenance.for: num  1 1 2 2 2 2 1 1 1 1 ...
#>  $ telephone                                               : Factor w/ 2 levels "none","yes, registered under the customers name": 2 1 1 1 1 2 1 2 1 1 ...
#>  $ foreign.worker                                          : Factor w/ 2 levels "yes","no": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ creditability                                           : Factor w/ 2 levels "bad","good": 2 1 2 2 1 2 2 2 2 1 ...

# summary of germancredit
lapply(germancredit, summary)

#> $status.of.existing.checking.account
#>                                             ... < 0 DM 
#>                                                    274 
#>                                      0 <= ... < 200 DM 
#>                                                    269 
#> ... >= 200 DM / salary assignments for at least 1 year 
#>                                                     63 
#>                                    no checking account 
#>                                                    394 
#> 
#> $duration.in.month
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     4.0    12.0    18.0    20.9    24.0    72.0 
#> 
#> $credit.history
#>                no credits taken/ all credits paid back duly 
#>                                                          40 
#>                     all credits at this bank paid back duly 
#>                                                          49 
#>                    existing credits paid back duly till now 
#>                                                         530 
#>                             delay in paying off in the past 
#>                                                          88 
#> critical account/ other credits existing (not at this bank) 
#>                                                         293 
#> 
#> $purpose
#>    Length     Class      Mode 
#>      1000 character character 
#> 
#> $credit.amount
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     250    1366    2320    3271    3972   18424 
#> 
#> $savings.account.and.bonds
#>                ... < 100 DM         100 <= ... < 500 DM 
#>                         603                         103 
#>        500 <= ... < 1000 DM              ... >= 1000 DM 
#>                          63                          48 
#> unknown/ no savings account 
#>                         183 
#> 
#> $present.employment.since
#>         unemployed       ... < 1 year 1 <= ... < 4 years 4 <= ... < 7 years 
#>                 62                172                339                174 
#>     ... >= 7 years 
#>                253 
#> 
#> $installment.rate.in.percentage.of.disposable.income
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   2.000   3.000   2.973   4.000   4.000 
#> 
#> $personal.status.and.sex
#>           male : divorced/separated female : divorced/separated/married 
#>                                   0                                 360 
#>                       male : single              male : married/widowed 
#>                                 548                                  92 
#>                     female : single 
#>                                   0 
#> 
#> $other.debtors.or.guarantors
#>         none co-applicant    guarantor 
#>          907           41           52 
#> 
#> $present.residence.since
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   2.000   3.000   2.845   4.000   4.000 
#> 
#> $property
#>                                          real estate 
#>                                                  282 
#>   building society savings agreement/ life insurance 
#>                                                  232 
#> car or other, not in attribute Savings account/bonds 
#>                                                  332 
#>                                unknown / no property 
#>                                                  154 
#> 
#> $age.in.years
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   19.00   27.00   33.00   35.55   42.00   75.00 
#> 
#> $other.installment.plans
#>   bank stores   none 
#>    139     47    814 
#> 
#> $housing
#>     rent      own for free 
#>      179      713      108 
#> 
#> $number.of.existing.credits.at.this.bank
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   1.000   1.000   1.407   2.000   4.000 
#> 
#> $job
#>                          unemployed/ unskilled - non-resident 
#>                                                            22 
#>                                          unskilled - resident 
#>                                                           200 
#>                                   skilled employee / official 
#>                                                           630 
#> management/ self-employed/ highly qualified employee/ officer 
#>                                                           148 
#> 
#> $number.of.people.being.liable.to.provide.maintenance.for
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   1.000   1.000   1.155   1.000   2.000 
#> 
#> $telephone
#>                                     none 
#>                                      596 
#> yes, registered under the customers name 
#>                                      404 
#> 
#> $foreign.worker
#> yes  no 
#> 963  37 
#> 
#> $creditability
#>  bad good 
#>  300  700 
#>